pyproject.toml
[tool.poetry]
name = "crypto-connect-back"
version = "0.1.0"
description = ""
authors = ["Anthony <anthonybeavermusic@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2"
django-money = "^1.3.1"
djangorestframework = "^3.12.4"
django-rest-auth = "^0.9.5"
djangorestframework-jwt = "^1.11.0"
whitenoise = "^5.2.0"
watchlist = "^0.0.7"


[tool.poetry.dev-dependencies]
autopep8 = "^1.5.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
